requirements.txt
requests
beautifulsoup4
python-dotenv
openai
slack_sdk
selenium
